Manually changed 'Quick3D.w3d' file
**[130] Treatment(1) - Main Icon**

Line 68: changed y-axis position on translate tag from 5.000000 to 3.000000. This is to align it with Floor 2.

**[130] Treatment(2) - Main Icon**

Line 72: changed y-axis position on translate tag from 5.000000 to 3.000000. This is to align it with Floor 2.

**[113] ProcessLoader(1) - Icon**

Line 14: changed y-axis position on translate tag from 1.123047 to 0.000000. This is to align it with Floor 1, which is also vertically aligned with the scene origin.