Fixed bug of certain parts not being created throughout simulation.
For instance, before 'Box - Entity (3)' was not being added to the simulation view because its update command did not occur at time zero and the SimulationCreateTag class was configured incorrectly to only respond to create commands at time zero. Now, this is no longer the case, and parts such as 'Box - Entity (3)' are being rendered to the simulation.

- Added Unit Tests for this functionality